requirements.txt
fastapi
uvicorn
boto3
langchain
langgraph
pydantic
python-dotenv
requests
openai
langchain-openai
langchain-core
langchain-community
# langchain-memory
# langchain-chains